Changes according mentor comments. Tabs refactoring. Renames ShowMore button handler.
