+++ src/components/show-more-button/show-more-button.jsx
@@ -2,21 +2,21 @@ import React from "react";
 import PropTypes from "prop-types";
 
 const ShowMoreButton = (props) => {
-  const {showMoreMovies} = props;
+  const {onShowMoreButtonClick} = props;
 
   return (
     <div className="catalog__more">
       <button
         className="catalog__button"
         type="button"
-        onClick={showMoreMovies}
+        onClick={onShowMoreButtonClick}
       >Show more</button>
     </div>
   );
 };
 
 ShowMoreButton.propTypes = {
-  showMoreMovies: PropTypes.func.isRequired
+  onShowMoreButtonClick: PropTypes.func.isRequired
 };
 
 export default ShowMoreButton;
